requirements.txt
fastapi
uvicorn
pydantic
openai
qdrant-client
sentence-transformers
python-dotenv
duckduckgo-search
bert-score
rouge-score
scikit-learn
torch
